+++ res/layout/list_fragment.xml
@@ -30,7 +30,7 @@
         android:layout_weight="1"
         android:footerDividersEnabled="false" > 
         
-        <third_parties.liaohuqiu.GridViewWithHeaderAndFooter
+        <third_parties.in.srain.cube.GridViewWithHeaderAndFooter
             android:id="@+id/list_root"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
