+++ src/third_parties/in/srain/cube/GridViewWithHeaderAndFooter.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package third_parties.liaohuqiu;
+package third_parties.in.srain.cube;
 
 
         import android.annotation.TargetApi;
